% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{abridge_df}
\alias{abridge_df}
\title{Summarize a data.frame using an opinionated -and perhaps controversial-
approach}
\usage{
abridge_df(df, file = NULL, ...)
}
\arguments{
\item{df}{data.frame-like object to summarize}

\item{file}{path to file to save the htmlwidget}

\item{...}{passed to htmlwidgets::saveWidget}
}
\value{
a DT datatables htmlwidget object
}
\description{
A one-function-call that produces a compact yet full of information
summary of a data.frame, presented in a DataTables (via \code{DT})
}
\examples{
abridge_df(mtcars)
abridge_df(ggplot2::diamonds)
nlsw88 <- haven::read_dta('http://www.stata-press.com/data/r15/nlsw88.dta')
abridge_df(nlsw88)
abridge_df(forcats::gss_cat)
}
