% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{are_paired}
\alias{are_paired}
\title{Check if the \code{...} columns of the \code{df} have a one-to-one relation}
\usage{
are_paired(df, ...)
}
\arguments{
\item{df}{a data.frame}

\item{...}{unquoted columns of the \code{df} to test. If empty, all columns used}
}
\value{
logical
}
\description{
Aim to test cases like:
\itemize{
\item A table with \code{person_id} and \code{birth_date}: each \code{person_id} must always
have the same \code{birth_date}. Calling \code{are_paired(df, person_id, birth_date)}
lets you test that. \code{birth_date}, of course, can be duplicated. Even
\code{person_id} could also be duplicated.
\item data with \code{region_name} and \code{region_code}
}
}
\details{
TODO: let's find a better name for this
TODO: check better algorithms
}
\examples{
are_paired(mtcars)
are_paired(mtcars, mpg, cyl)
}
