% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{attr}
\alias{attr}
\title{base::attr with a default value instead of NULL, and always exact}
\usage{
attr(x, which, default_value = "", exact = TRUE)
}
\arguments{
\item{x}{an object whose attributes are to be accessed.}

\item{which}{a non-empty character string specifying which attribute
    is to be accessed.}

\item{default_value}{value to return when object x has no attribute which}

\item{exact}{logical: should \code{which} be matched exactly?}
}
\description{
base::attr with a default value instead of NULL, and always exact
}
