% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{bar_just}
\alias{bar_just}
\title{Calculate hjust/vjust to align labels to the end of bars in ggplot2}
\usage{
bar_just(x, pivot = 0.5)
}
\arguments{
\item{x}{numeric}

\item{pivot}{numeric}
}
\value{
hjust or vjust
}
\description{
Calculate hjust/vjust to align labels to the end of bars in ggplot2
}
\examples{
library(ggplot2)
mtcars |>
tibble::rownames_to_column() |>
ggplot(aes(y = reorder(rowname, mpg), x = mpg, label = mpg)) +
geom_col(fill = "deepskyblue") +
geom_text(aes(hjust = bar_just(mpg)))
}
