% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{bar_just_facet}
\alias{bar_just_facet}
\title{Calculate hjust/vjust to align labels to the end of bars in ggplot2}
\usage{
bar_just_facet(x, facet, pivot = 0.5)
}
\arguments{
\item{x}{numeric}

\item{facet}{facets to group by}

\item{pivot}{numeric}
}
\value{
hjust or vjust
}
\description{
Calculate hjust/vjust to align labels to the end of bars in ggplot2
}
