% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogdown.R
\name{blogdown_stop_server}
\alias{blogdown_stop_server}
\title{Stop blogdown (hugo/jekyll) server}
\usage{
blogdown_stop_server()
}
\description{
Just a wrapper on \code{blogdown::stop_server()} to make it an addin
}
