% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipboard_readfrom}
\alias{clipboard_readfrom}
\title{Read from the clipboard to create a data.frame}
\usage{
clipboard_readfrom(sep = "\\t", header = TRUE, dec = ".", ...)
}
\arguments{
\item{sep}{use tab as default separator}

\item{header}{by default assume first line of data is column names}

\item{dec}{decimal separator to use}

\item{...}{passed to read.table}
}
\description{
Just for convenience to quickly input data from excel, plain text, etc.
by copy-'paste'
}
\seealso{
read.table()
}
\keyword{clipboard}
\keyword{read.table,}
