% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipboard_writeto}
\alias{clipboard_writeto}
\title{Copy an object to the clipboard}
\usage{
clipboard_writeto(eobject)
}
\arguments{
\item{eobject}{R object to copy to the clipboard}
}
\description{
It copies an object to the clipboard using tab as separator
(so as to paste in excel)
}
\seealso{
write.table()
}
\keyword{clipboard}
\keyword{write.table,}
