% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denstogram.R
\name{denstogram}
\alias{denstogram}
\title{Title}
\usage{
denstogram(
  data,
  xvar,
  fillvar = NULL,
  facets = NULL,
  grid_wrap = c("grid", "wrap"),
  scales = "fixed",
  facets_ncol = NULL,
  facets_nrow = NULL,
  trans = scales::identity_trans(),
  probs = c(0, 0.01, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99, 1),
  na.rm = TRUE,
  summary_ind = c("none", "line", "text", "both"),
  summ_fn = function(x) stats::median(x, na.rm = na.rm),
  plotly = TRUE
)
}
\arguments{
\item{data}{data}

\item{xvar}{var to plot}

\item{fillvar}{variable to fill denstogram by}

\item{facets}{variable to facet the denstogram by}

\item{grid_wrap}{whether to use a "grid" or "wrap" in facetting}

\item{scales}{passed to facets}

\item{facets_ncol}{cols for facets}

\item{facets_nrow}{rows for facets}

\item{trans}{transformation to apply from the \code{scales} package}

\item{probs}{quantiles to include in the plot}

\item{na.rm}{logical; if \code{TRUE}, missing values are removed
    from \code{x}. If \code{FALSE} any missing values cause an error.}

\item{plotly}{whether to convert it to an interactive plot using plotly}
}
\value{
ggplot object
}
\description{
Title
}
\examples{
denstogram(data = ggplot2::diamonds, xvar = price)
denstogram(data = ggplot2::diamonds, xvar = price, fillvar = color)
denstogram(data = ggplot2::diamonds, xvar = price, fillvar = cut)
denstogram(data = ggplot2::diamonds, xvar = price, facets = cut ~ .)
denstogram(data = ggplot2::diamonds, xvar = price, facets = cut ~ .,
           grid_wrap = "wrap")
denstogram(data = ggplot2::diamonds, xvar = price, facets = cut ~ color)
denstogram(data = ggplot2::diamonds, xvar = price, fillvar = cut,
           facets = color ~ .)
denstogram(data = ggplot2::diamonds, xvar = price, fillvar = cut,
           facets = color ~ ., grid_wrap = "wrap")
}
