% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Discretize a numeric vector}
\usage{
discretize(
  x,
  n = 5,
  method = break_methods(),
  labels_formatter = label_breaks_value,
  number_formatter = scales::number_format(big.mark = ","),
  ordered_result = TRUE
)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{integer indicating the number of intervals/bins desired}

\item{method}{a character of length 1, indicating the method to use. Possible
values are c("jenks", "isowidth", "quantile")}

\item{labels_formatter}{a function to produce the labels to pass to cut}

\item{number_formatter}{a function to format numbers, before building labels}

\item{ordered_result}{logical: should the result be an ordered factor?}
}
\value{
a vector as returned by cut, which is a factor, unless labels = FALSE
which results in an integer vector of level codes.
}
\description{
Wrap base::cut.default to make it convenient to use different binning
methods (see get_breaks), and custom functions for labels and number formats
}
\examples{
set.seed(1234567)

# quick example
set.seed(1234567)
discretize(runif(100))

# customize binning method
set.seed(1234567)
discretize(runif(100), method = "quantile")

# customize number format
set.seed(1234567)
discretize(runif(100), number_formatter = scales::percent)

# customize labels
set.seed(1234567)
discretize(runif(100), labels_formatter = label_breaks_interval)

set.seed(1234567)
discretize(runif(100), labels_formatter = label_breaks_cut)
}
