% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_explain_hook.R
\name{duckdb_explain_hook}
\alias{duckdb_explain_hook}
\title{Add chunk hook funtionality to print EXPLAIN results from DuckDB}
\usage{
duckdb_explain_hook(all = NULL)
}
\arguments{
\item{all}{whether to set all chunks as duckdb_explain}
}
\value{
function called for its side effects
}
\description{
ChatGPT generated explanaition: this code defines a custom knitr hook
(duckdb_explain) that modifies the rendering of SQL code chunks.
When executed before a code chunk, it formats the SQL query
explanation using HTML tags for preformatted code display.
When executed after a code chunk, it reverts the rendering
behavior to the default. This is useful for producing nicely
formatted query explanations in knitr documents.
}
\examples{
\dontrun{
# Simply put this line on the setup chunk of your .Rmd file
# and then set the chunk option `duckdb_explain` to TRUE when
# you want to override standard printing of sql chunks
efun::duckdb_explain_hook()
}
}
