% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfgram.R
\name{ecdfgram}
\alias{ecdfgram}
\title{Plot Empirical Cummulative Distribution Function for xvar}
\usage{
ecdfgram(
  data,
  xvar,
  grpvar = NULL,
  trans = scales::identity_trans(),
  xaxis_n_breaks = 7,
  xaxis_title = "",
  yaxis_title = "Cummulative Frequency",
  xaxis_format_args = list()
)
}
\arguments{
\item{data}{data to plot}

\item{xvar}{xvar to plot}

\item{grpvar}{var to group the data by}

\item{trans}{transformation to apply from the \code{scales} package}

\item{xaxis_n_breaks}{number of breaks on the x_axis}

\item{xaxis_title}{x axis title}

\item{yaxis_title}{y axis title}

\item{xaxis_format_args}{#TODO}
}
\value{
a plotly object
}
\description{
Plot Empirical Cummulative Distribution Function for xvar
}
\examples{
ecdfgram(ggplot2::diamonds, price, color)
}
