% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efun-package.R
\docType{package}
\name{efun-package}
\alias{efun}
\alias{efun-package}
\title{efun: Miscellaneous Functions by E}
\description{
A package where I put functions that, at the time, did not belong somewhere else. A bunch of those functions get refactored, but some of them stick around as utility functions.
}
\author{
\strong{Maintainer}: Eduardo Alfonso-Sierra \email{edalfon@gmail.com} (\href{https://orcid.org/0000-0001-6430-5135}{ORCID}) [copyright holder]

}
\keyword{internal}
