% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{ellipt_colnames}
\alias{ellipt_colnames}
\title{"Delete by ellipsis" column names, putting the original
column names as a tooltip}
\usage{
ellipt_colnames(data, ellipsis = "", min_width = 3)
}
\arguments{
\item{data}{data whose columns will be renamed}

\item{ellipsis}{indicator to append to the column names to inducate trunc}

\item{min_width}{minimum with to guarantee for the column names}
}
\value{
data with new column names
}
\description{
"Delete by ellipsis" column names, putting the original
column names as a tooltip
}
