% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-file.R
\name{file_glimpse}
\alias{file_glimpse}
\title{Glimpse first and last n/2 lines of a file}
\usage{
file_glimpse(file_path, n = 14)
}
\arguments{
\item{file_path}{file path (character)}

\item{n}{number of lines (integer)}
}
\value{
a character vector (length n; if one_string = TRUE, length 1)
}
\description{
And print the lines (using cat)
}
