% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-file.R
\name{file_tail}
\alias{file_tail}
\title{Glimpse last n lines of a file}
\usage{
file_tail(file_path, n = 10, one_string = TRUE)
}
\arguments{
\item{file_path}{file path (character)}

\item{n}{number of lines (integer)}

\item{one_string}{logical, if TRUE (default) it collapses all lines in
a singe string}
}
\value{
a character vector (length n; if one_string = TRUE, length 1)
}
\description{
It uses powershell (currently a Windows function only)
}
