% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicates.R
\name{filter_duplicates}
\alias{filter_duplicates}
\title{Filter duplicated row by the columns indicated in by}
\usage{
filter_duplicates(data, by)
}
\arguments{
\item{data}{a data.frame-like}

\item{by}{a character vector indicating the columns to consider to identify
duplicated rows}
}
\value{
a data.table, subset of the original data, containing only
duplicated rows by the columns indicated in \code{by}, and including an
additional column .ndups with the number of duplicated rows for each group.
The returned data are sorted by the columns, to be able to inspect
the duplicated together
}
\description{
Filter duplicated row by the columns indicated in by
}
\examples{
filter_duplicates(ggplot2::diamonds, by = c("carat", "cut", "price"))
}
