% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicates.R
\name{filter_duplicates_dplyr}
\alias{filter_duplicates_dplyr}
\title{Filter duplicated row by the columns ...}
\usage{
filter_duplicates_dplyr(data, ...)
}
\arguments{
\item{data}{a data.frame-like}

\item{...}{columns (unquoted) to consider to identify duplicated rows}
}
\value{
a tibble, subset of data, containing only duplicated rows by the
columns indicated by ..., and including an additional column .ndups with the
number of duplicated rows for each group
The returned data are sorted by the columns, to be able to inspect
the duplicated together
}
\description{
Filter duplicated row by the columns ...
}
\examples{
filter_duplicates_dplyr(ggplot2::diamonds, carat, cut, price)
}
