% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_non_null.R
\name{first_non_null}
\alias{first_non_null}
\title{Return the first non-null argument}
\usage{
first_non_null(...)
}
\arguments{
\item{...}{One or more arguments of any type.}
}
\value{
The first non-null argument or \code{NULL} if all arguments are null.
}
\description{
This is an old coalesce functions for its arguments. It was actually
called coalesce, but now dplyr has also one of those. So let's just
The \code{first_non_null} function takes any number of arguments and returns the
first argument that is not null (i.e., \code{NA}). If all arguments are null, the
function returns \code{NULL}.
}
