% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{get_breaks}
\alias{get_breaks}
\title{Calculate breaks to discretize a numeric vector, using different methods}
\usage{
get_breaks(x, n = 5, method = break_methods())
}
\arguments{
\item{x}{a numeric vector}

\item{n}{integer indicating the number of intervals/bins desired}

\item{method}{a character of length 1, indicating the method to use. Possible
values are c("jenks", "isowidth", "quantile")}
}
\value{
a numeric vector of length <= n + 1.
}
\description{
Calculate breaks to discretize a numeric vector, using different methods
}
\examples{
get_breaks(1:100, 5)
}
