% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%#_\%}
\alias{\%#_\%}
\title{Function that defines the operator \%#\% to concatenate strings}
\usage{
x \%#_\% y
}
\arguments{
\item{x}{character object, to be converted to character vector
(passing it to paste)}

\item{y}{character object, to be converted to character vector
(passing it to paste)}
}
\value{
a character vector
}
\description{
This simple calls paste with sep = ""
}
\keyword{paste}
