% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%#\%}
\alias{\%#\%}
\alias{\%# \%}
\title{Function that defines the operator \%#\% to concatenate strings}
\usage{
x \%#\% y

x \%# \% y
}
\arguments{
\item{x}{character object, to be converted to character vector
(passing it to paste)}

\item{y}{character object, to be converted to character vector
(passing it to paste)}
}
\value{
a character vector

a character vector
}
\description{
This simple calls paste with sep = ""

This simple calls paste with sep = " "
}
\keyword{paste}
