% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{is_invalid_number}
\alias{is_invalid_number}
\title{Test if x is NA, NaN, Inf, -Inf, NULL}
\usage{
is_invalid_number(x)
}
\arguments{
\item{x}{a vector}
}
\value{
boolean
}
\description{
Can be helpful as a workaround for gotchas such as:
is.numeric(NaN)
is.numeric(Inf)
is.numeric(NA)
is.numeric(NULL)
}
\examples{
is_invalid_number(c(NA, Inf, mean(Inf, -Inf)))
}
