% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{is_unique}
\alias{is_unique}
\title{Check if a variable is unique, i.e., there are as many unique values in the
object as the total number of values}
\usage{
is_unique(obj)
}
\arguments{
\item{obj}{a vector}
}
\description{
Check if a variable is unique, i.e., there are as many unique values in the
object as the total number of values
}
\examples{
is_unique(1:7)
}
\seealso{
unique length
}
