% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{label_breaks_cut}
\alias{label_breaks_cut}
\title{Label breaks as base::cut would do}
\usage{
label_breaks_cut(breaks)
}
\arguments{
\item{breaks}{a vector}
}
\value{
a vector of length, length(breaks) - 1
}
\description{
Label breaks as base::cut would do
}
