% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{label_breaks_interval}
\alias{label_breaks_interval}
\title{Label breaks as an interval}
\usage{
label_breaks_interval(breaks)
}
\arguments{
\item{breaks}{a vector}
}
\value{
a vector of length, length(breaks) - 1
}
\description{
Label breaks as an interval
}
