% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{label_breaks_value}
\alias{label_breaks_value}
\title{Label breaks using only point values (discarding lowest)}
\usage{
label_breaks_value(breaks)
}
\arguments{
\item{breaks}{a vector}
}
\value{
a vector of length, length(breaks) - 1
}
\description{
Label breaks using only point values (discarding lowest)
}
