% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{make_mix}
\alias{make_mix}
\title{Where we make a mess}
\usage{
make_mix(stat_var, top_key, top_count, top_count_p, top_k)
}
\arguments{
\item{stat_var}{stat var from the summary df, such as mean, sd, q25, etc.}

\item{top_key}{value of the top X}

\item{top_count}{count of the top X}

\item{top_count_p}{percentage of the top X}

\item{top_k}{X, a number indicating the top}
}
\description{
the part where abridge_df is most strongly opinionated is in the mix of
stats for numeric and non-numeric variables in the same column.
Basically, a bunch of columns where we show stats (e.g. mean) for numeric
variables, and top X value and counts for non-numeric variables.
So in this function we pass the relevant variables from the summary df
and it makes this conditional
}
\keyword{internal}
