% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{memory_use}
\alias{memory_use}
\title{Print memory use (total and each object in the global environment)}
\usage{
memory_use()
}
\description{
By default, using Mb as units. See ?memory.size()
Also, this is for windows only
}
\seealso{
object.size, memory.size, memory.limit
}
