% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modemfv.R
\name{modemfv_vctrs}
\alias{modemfv_vctrs}
\title{Find the mode in a vector x}
\usage{
modemfv_vctrs(x)
}
\arguments{
\item{x}{a vector}
}
\value{
object of length 1, the same type as x
}
\description{
Find the mode in a vector x
}
\examples{
modemfv_vctrs(ggplot2::diamonds$cut)
}
