% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{normalize_text}
\alias{normalize_text}
\title{Normalize text, removing spaces, common punctuation and non-ascii characters}
\usage{
normalize_text(text)
}
\arguments{
\item{text}{A character vector}
}
\value{
A transformed character vector
}
\description{
Normalize text, removing spaces, common punctuation and non-ascii characters
}
\examples{
normalize_text("§Bogotá, D.C.")
}
