% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_file.R
\name{open_file}
\alias{open_file}
\title{Open a file or URL in the system's default application}
\usage{
open_file(file_path)
}
\arguments{
\item{file_path}{a vector}
}
\value{
see ?system2
}
\description{
Open a file or URL in the system's default application
}
