% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_annotation.R
\name{parse_annotations}
\alias{parse_annotations}
\title{Parse an hierarchical annotation into a data.frame}
\usage{
parse_annotations(notes, pattern = "#+")
}
\arguments{
\item{notes}{a character vector}

\item{pattern}{a length 1 character vector, with the pattern to parse the
annotation. This is passed to \code{stringr::str_split} and
\code{stringr::str_extract_all}. The default value matches 1 or more hashtags (#)
and succesive hashtags indicate the level of the title/code}
}
\value{
a list of data.frame
}
\description{
Parse an hierarchical annotation into a data.frame
}
\seealso{
parse_single_annotation
}
