% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_annotation.R
\name{parse_single_annotation}
\alias{parse_single_annotation}
\title{Parse an hierarchical annotation into a data.frame}
\usage{
parse_single_annotation(note, pattern = "#+")
}
\arguments{
\item{note}{a length 1 character vector, with an annotation to be parsed}

\item{pattern}{a length 1 character vector, with the pattern to parse the
annotation. This is passed to \code{stringr::str_split} and
\code{stringr::str_extract_all}. The default value matches 1 or more hashtags (#)
and succesive hashtags indicate the level of the title/code}
}
\value{
a data.frame
}
\description{
This function converts a markdown-style hierarchical annotation
into a data.frame with from -> to columns. For example a text like this:#'
#Heading 1
#Heading 2
##Heading 2.1
##Heading 2.2
#Heading 3
would be converted to a data.frame with all the links
(e.g. Heading 2 -> Heading 2.1)
}
\examples{
parse_single_annotation(
  "#1 ##1.1 ##1.2 ###1.2.1 ###1.2.2 ##1.3 #2 ##2.1 ###2.1.1"
)
}
