% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{pg_create_foreign_table}
\alias{pg_create_foreign_table}
\title{Create a foreign table in Postgres, to directly access data in a text file}
\usage{
pg_create_foreign_table(
  con,
  file_path,
  table,
  sep = ";",
  nrows = 10000,
  header = TRUE,
  na.strings = c("", "NA", "NULL"),
  colClasses = NULL,
  encoding = "UTF-8",
  ...,
  drop_table = FALSE,
  cascade = FALSE,
  if_not_exists = FALSE,
  create_foreign_server = TRUE,
  execute = FALSE,
  quote = "\\b",
  escape = quote,
  null = "NULL"
)
}
\arguments{
\item{con}{A database connection.}

\item{file_path}{path to a text file (csv, tab-delimited, etc.)}

\item{table}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{sep}{ The separator between columns. Defaults to the character in the set \code{[,\\t |;:]} that separates the sample of rows into the most number of lines with the same number of fields. Use \code{NULL} or \code{""} to specify no separator; i.e. each line a single character column like \code{base::readLines} does.}

\item{nrows}{ The maximum number of rows to read. Unlike \code{read.table}, you do not need to set this to an estimate of the number of rows in the file for better speed because that is already automatically determined by \code{fread} almost instantly using the large sample of lines. \code{nrows=0} returns the column names and typed empty columns determined by the large sample; useful for a dry run of a large file or to quickly check format consistency of a set of files before starting to read any of them. }

\item{header}{ Does the first data line contain column names? Defaults according to whether every non-empty field on the first data line is type character. If so, or TRUE is supplied, any empty column names are given a default name. }

\item{na.strings}{ A character vector of strings which are to be interpreted as \code{NA} values. By default, \code{",,"} for columns of all types, including type \code{character} is read as \code{NA} for consistency. \code{,"",} is unambiguous and read as an empty string. To read \code{,NA,} as \code{NA}, set \code{na.strings="NA"}. To read \code{,,} as blank string \code{""}, set \code{na.strings=NULL}. When they occur in the file, the strings in \code{na.strings} should not appear quoted since that is how the string literal \code{,"NA",} is distinguished from \code{,NA,}, for example, when \code{na.strings="NA"}. }

\item{colClasses}{ As in \code{\link[utils:read.table]{utils::read.csv}}; i.e., an unnamed vector of types corresponding to the columns in the file, or a named vector specifying types for a subset of the columns by name. The default, \code{NULL} means types are inferred from the data in the file. Further, \code{data.table} supports a named \code{list} of vectors of column names \emph{or numbers} where the \code{list} names are the class names; see examples. The \code{list} form makes it easier to set a batch of columns to be a particular class. When column numbers are used in the \code{list} form, they refer to the column number in the file not the column number after \code{select} or \code{drop} has been applied.
    If type coercion results in an error, introduces \code{NA}s, or would result in loss of accuracy, the coercion attempt is aborted for that column with warning and the column's type is left unchanged. If you really desire data loss (e.g. reading \code{3.14} as \code{integer}) you have to truncate such columns afterwards yourself explicitly so that this is clear to future readers of your code.
  }

\item{encoding}{ default is \code{"unknown"}. Other possible options are \code{"UTF-8"} and \code{"Latin-1"}.  Note: it is not used to re-encode the input, rather enables handling of encoded strings in their native encoding. }

\item{...}{Other arguments used by individual methods.}

\item{drop_table}{logical, whether to drop the table before creating it}

\item{cascade}{logical, whether to add CASCADE to the DROP statement}

\item{if_not_exists}{logical, to add IF NOT EXISTS to the query}

\item{create_foreign_server}{logical, to create the extension and server
for the foreign data wrapper (\url{https://www.postgresql.org/docs/10/file-fdw.html})}

\item{execute}{logical, whether to execute the query using \code{con}}

\item{quote}{ By default (\code{"\\""}), if a field starts with a double quote, \code{fread} handles embedded quotes robustly as explained under \code{Details}. If it fails, then another attempt is made to read the field \emph{as is}, i.e., as if quotes are disabled. By setting \code{quote=""}, the field is always read as if quotes are disabled. It is not expected to ever need to pass anything other than \\"\\" to quote; i.e., to turn it off. }

\item{escape}{"character that should appear before a data character that
matches the QUOTE value"}

\item{null}{"Specifies the string that represents a null value."}
}
\value{
invisibly returns a character vector (length 1) with the sql
statement
}
\description{
The use case for this function is as follows:
You get a large file (csv, tab-delimited) and would rather wrangle such
data in postgres. Instead of go ahead and copy all the data into a table,
let's create a foreign table that you can query just like a regular table.
}
\details{
Some notes:
\itemize{
\item The file must be accesible to the postgres server. This function tries to
edit file's privileges to make sure is readable from the server.
However, if the file is in a network share, that may not be enough because
the server should be able to access the network resource. It is likely
the default user running the server cannot access network shared resources,
so the workaround is to change the user running the server
}

See Postgres documentation:
\itemize{
\item \href{https://www.postgresql.org/docs/10/sql-createforeigntable.html}{Foreign table https://www.postgresql.org/docs/10/sql-createforeigntable .html}
\item \href{https://www.postgresql.org/docs/10/file-fdw.html}{Foreign-data wrapper for files https://www.postgresql.org/docs/10/file- fdw.html}
\item \href{https://www.postgresql.org/docs/10/sql-copy.html}{Options documented in COPY https://www.postgresql.org/docs/10/sql-copy .html}
}
}
