% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{pg_exists_table}
\alias{pg_exists_table}
\title{Check if \code{table_name} exists in the Postgres database of given \code{con}}
\usage{
pg_exists_table(con, table_name)
}
\arguments{
\item{con}{DBI connection to the database}

\item{table_name}{character with the table name you want to check}
}
\value{
boolean
}
\description{
Uses Postgres' \code{to_regclass}, taking advantage of the fact that it will
"return null rather than throwing an error if the name is not found"
https://www.postgresql.org/docs/10/functions-info.html
This is probably the fastest way to test if a table (relation) exists.
Alternatives would be querying \code{pg_tables} or \code{pg_catalog}
}
