% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{pg_running_queries}
\alias{pg_running_queries}
\title{List running queries in Postgres}
\usage{
pg_running_queries(con)
}
\arguments{
\item{con}{pg connection}
}
\value{
a tibble
}
\description{
List running queries in Postgres
}
