% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{pg_total_relation_size}
\alias{pg_total_relation_size}
\title{Query table's size in the Postgres database of given \code{con}}
\usage{
pg_total_relation_size(con, table_name, pretty = TRUE)
}
\arguments{
\item{con}{DBI connection to the database}

\item{table_name}{character with the table name you want to check}

\item{pretty}{logical, if TRUE use \code{pg_size_pretty}}
}
\value{
character if pretty=TRUE, numeric otherwise
}
\description{
Uses Postgres' \code{pg_total_relation_size} to get the value, and
\code{pg_size_pretty} to try and make it human readable directly
}
