% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{phased_left_join}
\alias{phased_left_join}
\title{Left join lhs and rhs, using a phased approach}
\usage{
phased_left_join(
  lhs,
  rhs,
  phased_by,
  exact_by = NULL,
  drop_join_vars = TRUE,
  quasi_fun = normalize_text,
  suffix = c(".x", ".y")
)
}
\arguments{
\item{lhs}{A data.frame-like}

\item{rhs}{A data.frame-like}

\item{phased_by}{A character vector of variables to join by, using a named
vector to join by different variables from \code{lhs} and \code{rhs} (à la \code{dplyr}.
See ?dplyr::join). The variables indicated here will be subject to the phased
approach.}

\item{exact_by}{An optional character vector of variables to join by, using
always an exact match.}

\item{drop_join_vars}{Whether to drop auxiliary variables for the match
(e.g. transformed variables). Defaults to TRUE, but for debugging may be
useful to set to FALSE.}

\item{quasi_fun}{The function to apply to \code{phased_by} variables, for the
quasi-exact match. Defaults to efun::normalize_text}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and
\code{y}, these suffixes will be added to the output to disambiguate them.
Should be a character vector of length 2.}
}
\value{
a joined data-frame
}
\description{
We use three 'phases' to left-join \code{lhs} and \code{rhs}
\itemize{
\item \strong{Exact match}: well, exact match both on \code{exact_by} and \code{phased_by}
\item \strong{Quasi-exact match}: transform \code{phased_by} vars on \code{lhs} and \code{rhs} using
\code{quasi_fun} and then matches on the transformed variables. Default for
\code{quasi_fun} is \code{efun::normalize_text}, which removes spaces, dots, commas
and non-ascii characters to avoid encoding issues
\item \strong{Fuzzy match using a two-way 'contains' approach}: this is powered by
\code{fuzzyjoin::fuzzy_left_join} using as matching function
\code{match_fun = ~ stringr::str_detect(.x, .y) | stringr::str_detect(.y, .x)}
}
}
\details{
We apply those phases in that order, and every phase works only on the
unmatched rows from previous phases.
}
