% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_df}
\alias{summarize_df}
\title{Summarize a data.frame by applying \code{summarize_obj} to each column}
\usage{
summarize_df(df)
}
\arguments{
\item{df}{data.frame object to summarize}
}
\value{
a data.frame summarizing the original data.frame with information
such as length (number of elements), number of na, number of unique elements,
sum, min, max, mean, median, percentiles, most frequent values
}
\description{
Iterates through all the objects of a dataset, call \code{summarize_obj} on each,
and aggregates the results in a data.frame. This one is not meant to be
called directly, because it returns a data.frame full of information but
not really visual friendly. But we export it anyway, in case it is useful
at some point.
}
\examples{
summarize_df(mtcars)
}
