% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_obj}
\alias{summarize_obj}
\title{Summarize an object, typically a vector (a column from a data.frame)}
\usage{
summarize_obj(obj, n_top = 7)
}
\arguments{
\item{obj}{R object to summarize}

\item{n_top}{Number of top unique elements to include in the summary}
}
\value{
a one-row tibble with information about the object such as class,
length (number of elements), number of na, number of unique elements, sum,
min, max, mean, median, percentiles, most frequent values
}
\description{
Summarize an object (typically, a variable -column- in a
data.frame). This function is intended to be called by \code{summarize_df}
}
\examples{
summarize_obj(letters)
summarize_obj(runif(100))
}
