% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{tab}
\alias{tab}
\title{One-way tabulate table}
\usage{
tab(vctr, .drop = FALSE)
}
\arguments{
\item{vctr}{a vector to tabulate}

\item{.drop}{For \code{count()}: if \code{FALSE} will include counts for empty groups
(i.e. for levels of factors that don't exist in the data). Deprecated in
\code{add_count()} since it didn't actually affect the output.}
}
\value{
tibble with columns value, n, perc
}
\description{
Returned as tibble, with percentages
}
\examples{
vctr <- sample(LETTERS[1:5], 1000, TRUE)
tab(vctr)
# Includes NAs count
vctr <- sample(c(LETTERS[1:5], NA), 1000, TRUE)
tab(vctr)
# For factors, by default keeps levels not present in data
vctr <- factor(
  x = sample(c(LETTERS[1:5], NA), 1000, TRUE),
  levels = c(LETTERS[1:5], "Levels", "not present", "in data")
)
tab(vctr)
}
