% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{time_chunk}
\alias{time_chunk}
\title{Time \code{knitr} chunks, by default, all of them}
\usage{
time_chunk(all = TRUE)
}
\arguments{
\item{all}{whether to time all chunks by default (set the default value for
chunk option \code{time_it})}
}
\description{
This function takes knitr's official example on using chunk hooks to
report how much time each chunk takes to run:
\url{https://bookdown.org/yihui/rmarkdown-cookbook/time-chunk.html}
}
\details{
And just wrap it in a function so you can simply write \code{efun::time_chunk()}
in your top chunk, instead of the whole example.
}
\examples{
\dontrun{
# Simply put this line on the top chunk of your .Rmd file
efun::time_chunk(all = TRUE)
}
}
