% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{timestamp_it}
\alias{timestamp_it}
\title{Append a timestamp to \code{x}, typically a path}
\usage{
timestamp_it(x, is_path = TRUE)
}
\arguments{
\item{x}{a character or an object coercible to character}

\item{is_path}{logical to indicate if x is a path, then it removes extension
before appending the timestamp}
}
\value{
a character appending a timestamp to x
}
\description{
This function just appends a timestamp (format YYYYMMDDHHMMSS), separated by
an underscore (_) to whatever \code{x} is (so, \code{x} is coerced to character).
By default, it assumes \code{x} is a path and then adds the timestamp before
the last extension (if there is any).
}
\examples{
# typical usage
timestamp_it("this/path/may/not/exist/")
timestamp_it("this/path/may/not/exist.txt")
# weird usage
timestamp_it(mtcars, FALSE)
}
