% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{visualize_summary}
\alias{visualize_summary}
\title{Put the summary in a DT for compact, useful and hopefully pretty
visualization}
\usage{
visualize_summary(summary_df)
}
\arguments{
\item{summary_df}{}
}
\value{
a DT datatables htmlwidget object
}
\description{
This function takes as argument a summary data.frame as returned by
\code{summarize_df}, which contains all the information extracted from the
original data.frame to be presented in the DT summary.
}
\details{
We take advantage of DT to:
\itemize{
\item style columns with color bars to easilly pinpoint variables with many/few
NAs and high/low cardinality variables
\item Include tooltips in several places to put more information available,
without using space on the table (not using the datatables ellipsis
plugin, though. Some cases would be kind of hacky, so it ends up being
cleaner just changing the content of the data directly to include tooltips)
\item of course, enable sorting by the different stats
\item and we use \code{sparkline} to include inline histograms for each variable#'
}
}
\keyword{internal}
