% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{get_stats}
\alias{get_stats}
\title{Calculate several stats for obj}
\usage{
get_stats(obj)
}
\arguments{
\item{obj}{}
}
\value{
a named vector of stats
}
\description{
Using S3 methods to dispatch get_stats according to the obj class
}
\keyword{internal}
